+++ src/Factories/TemplatedStr.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Dhii\Services\Factories;
+
+class TemplatedStr extends StringService
+{
+}